+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/source/TagAutocomplete.java
@@ -16,31 +16,31 @@
  *
  */
 
-package org.apache.skywalking.oap.server.receiver.zipkin;
+package org.apache.skywalking.oap.server.core.source;
 
 import lombok.Getter;
 import lombok.Setter;
-import org.apache.skywalking.oap.server.core.Const;
-import org.apache.skywalking.oap.server.library.module.ModuleConfig;
 
-import java.util.ArrayList;
-import java.util.List;
+public abstract class TagAutocomplete extends Source {
 
-@Setter
-@Getter
-public class ZipkinReceiverConfig extends ModuleConfig {
-    private String restHost;
-    private int restPort;
-    private String restContextPath;
-    private int restMaxThreads = 200;
-    private long restIdleTimeOut = 30000;
-    private int restAcceptorPriorityDelta = 0;
-    private int restAcceptQueueSize = 0;
-    private List<String> instanceNameRule = new ArrayList<>();
-    private String searchableTracesTags = DEFAULT_SEARCHABLE_TAG_KEYS;
+    @Override
+    public int scope() {
+        return DefaultScopeDefine.TRACE_TAG_AUTOCOMPLETE;
+    }
 
-    private static final String DEFAULT_SEARCHABLE_TAG_KEYS = String.join(
-        Const.COMMA,
-        "http.method"
-    );
+    @Override
+    public String getEntityId() {
+        return tag;
+    }
+
+    @Setter
+    @Getter
+    private String tag;
+    @Setter
+    @Getter
+    private String tagKey;
+    @Setter
+    @Getter
+    private String tagValue;
+
 }