+++ oap-server/server-receiver-plugin/zipkin-receiver-plugin/src/main/java/org/apache/skywalking/oap/server/receiver/zipkin/trace/SpanForward.java
@@ -18,10 +18,13 @@
 
 package org.apache.skywalking.oap.server.receiver.zipkin.trace;
 
+import java.util.Arrays;
 import java.util.List;
 
 import lombok.RequiredArgsConstructor;
+import org.apache.skywalking.oap.server.core.Const;
 import org.apache.skywalking.oap.server.core.analysis.Layer;
+import org.apache.skywalking.oap.server.core.source.TraceTagAutocomplete;
 import org.apache.skywalking.oap.server.library.util.StringUtil;
 import org.apache.skywalking.oap.server.core.analysis.IDManager;
 import org.apache.skywalking.oap.server.core.analysis.TimeBucket;
@@ -86,8 +89,13 @@ public class SpanForward {
             zipkinSpan.setLatency((int) latency);
             zipkinSpan.setDataBinary(SpanBytesEncoder.PROTO3.encode(span));
 
+            List<String> searchTagKeys = Arrays.asList(config.getSearchableTracesTags().split(Const.COMMA));
             span.tags().forEach((key, value) -> {
-                zipkinSpan.getTags().add(key + "=" + value);
+                if (searchTagKeys.contains(key)) {
+                    String tagString = key + "=" + value;
+                    zipkinSpan.getTags().add(tagString);
+                    addAutocompleteTags(minuteTimeBucket, key, value, tagString);
+                }
             });
 
             receiver.receive(zipkinSpan);
@@ -100,6 +108,15 @@ public class SpanForward {
             serviceMeta.setLayer(Layer.GENERAL);
             receiver.receive(serviceMeta);
         });
+    }
+
+    private void addAutocompleteTags(final long minuteTimeBucket, final String key, final String value, final String tagString) {
+        TraceTagAutocomplete tagAutocomplete = new TraceTagAutocomplete();
+        tagAutocomplete.setTag(tagString);
+        tagAutocomplete.setTagKey(key);
+        tagAutocomplete.setTagValue(value);
+        tagAutocomplete.setTimeBucket(minuteTimeBucket);
+        receiver.receive(tagAutocomplete);
     }
 
     private String getServiceInstanceName(Span span) {