+++ oap-server/server-storage-plugin/storage-iotdb-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/iotdb/IoTDBTableMetaInfo.java
@@ -85,6 +85,9 @@ public class IoTDBTableMetaInfo {
         if (storageAndIndexMap.containsValue(IoTDBIndexes.AGENT_ID_INX)) {
             indexes.add(IoTDBIndexes.AGENT_ID_INX);
         }
+        if (storageAndIndexMap.containsValue(IoTDBIndexes.AUTOCOMPLETE_TAG_KEY)) {
+            indexes.add(IoTDBIndexes.AUTOCOMPLETE_TAG_KEY);
+        }
 
         final IoTDBTableMetaInfo tableMetaInfo = IoTDBTableMetaInfo.builder()
                                                                    .model(model)